max size training works. the size is 2016 x 1344 Z (!)
it's basically defined in `icdar.py`, and `generator_maxsize()`, which
is called in `generator()`
I added the new flag of `--prep`, which means preprocessing method.
E.g. to use the max size, add `--prep max`

Here is the command i use for train. The max batch size per gpu is 2.
put -1 for original size so that we don't accidentally use  it somewhere

```
python multigpu_train.py --checkpoint_path=../experiments/east-char
--training_data_path=../data/east_char_train
--geometry=RBOX
--learning_rate=0.0001  --input_size=-1  --text_scale=-1
--num_readers=24 --pretrained_model_path=./resnet_v1_50.ckpt
--batch_size_per_gpu 2
--gpu_list=6 --prep max
```